tests/wpt/tests/html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function-src.html

<!doctype html>
<meta charset=utf-8>
<title>Set src from a function called from a parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe src="support/set-parent-src.html"></iframe>
<script>
async_test(function() {
  onload = this.step_func(function() {
    var fr = document.querySelector("iframe")
    fr.contentWindow.go()
    fr.onload = this.step_func_done(function() {
      assert_equals(fr.contentDocument.referrer, document.URL)
    })
  })
})
</script>
